+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 # Changelog
 
 # Todo
+- jsbaton - migrate function dbExec() to new unified jsbaton
 - jenks - replace class-based api with static-function
 - wasm - replace sqljs-api with sqlmath-api in wasm
 - sqlmath - add sql-extension carrayblob()
 - none
 
 # v2022.6.1-beta
+- jsbaton - migrate function dbTableInsert() to new unified jsbaton
 - jsbaton - migrate functions dbClose(), dbMemoryLoadOrSave(), dbNoop(), dbOpen() to new, streamlined jsbaton2
 - add js-function jsbatonPushValue() to dynamically push value to jsbaton
 - napi - decouple c-functions dbClose(), dbExec(), dbMemoryLoadOrSave(), dbNoop(), dbOpen(), dbTableInsert() from napi